+++ R/ce_plot.R
@@ -4,15 +4,25 @@
 #' @param ref_col Which cutpoint method to use as the reference strategy
 #' when calculating the incremental net monetary benefit. Often sensible to use
 #' a "all" or "none" approach for this.
-#' @param wtp A \code{numeric}. The willingness to pay value used to create a
-#' cost-effectiveness plane on the plot (if \code{show_wtp = TRUE}).
+#' @param wtp A \code{numeric}. The willingness to pay (WTP) value used to create a
+#' WTP threshold line on the plot (if \code{show_wtp = TRUE}). Defaults to the
+#' WTP stored in the \code{predictNMBsim} object.
 #' @param show_wtp A \code{logical}. Whether or not to show the willingness to
-#' pay plane.
+#' pay threshold.
 #' @param methods_order The order (within the legend) to display the
 #' cutpoint methods.
 #' @param rename_vector A named vector for renaming the methods in the summary.
 #' The values of the vector are the default names and the names given are the
 #' desired names in the output.
+#' @param shape The \code{shape} used for \code{ggplot2::geom_point()}.
+#' Defaults to 21 (hollow circles). If \code{shape = "method"} or
+#' \code{shape = "cost-effective"} (only applicable when \code{show_wtp = TRUE})
+#' , then the shape will be mapped to that aesthetic.
+#' @param wtp_linetype The \code{linetype} used for \code{ggplot2::geom_abline()}
+#' when making the WTP. Defaults to \code{"dashed"}.
+#' @param add_prop_ce Whether to append the proportion of simulations for that
+#' method which were cost-effective (beneath the WTP threshold)
+#' to their labels in the legend. Only applicable when \code{show_wtp = TRUE}.
 #' @param ... Additional (unused) arguments.
 #'
 #' @details
@@ -45,6 +55,9 @@ ce_plot <- function(object,
                     show_wtp = TRUE,
                     methods_order = NULL,
                     rename_vector,
+                    shape = 21,
+                    wtp_linetype = "dashed",
+                    add_prop_ce = FALSE,
                     ...) {
   UseMethod("ce_plot")
 }
@@ -56,17 +69,24 @@ ce_plot <- function(object,
 #' @param ref_col Which cutpoint method to use as the reference strategy
 #' when calculating the incremental net monetary benefit. Often sensible to use
 #' a "all" or "none" approach for this.
-#' @param wtp A \code{numeric}. The willingness to pay value used to create a
-#' cost-effectiveness plane on the plot (if \code{show_wtp = TRUE}).
-#' @param show_wtp A \code{logical}. Whether or not to show the willingness to
-#' pay plane.
+#' @param wtp A \code{numeric}. The willingness to pay (WTP) value used to create a
+#' WTP threshold line on the plot (if \code{show_wtp = TRUE}). Defaults to the
+#' WTP stored in the \code{predictNMBsim} object.
+#' @param show_wtp A \code{logical}. Whether or not to show the WTP threshold.
 #' @param methods_order The order (within the legend) to display the
 #' cutpoint methods.
 #' @param rename_vector A named vector for renaming the methods in the summary.
 #' The values of the vector are the default names and the names given are the
 #' desired names in the output.
 #' @param shape The \code{shape} used for \code{ggplot2::geom_point()}.
-#' Defaults to 21 (hollow circles).
+#' Defaults to 21 (hollow circles). If \code{shape = "method"} or
+#' \code{shape = "cost-effective"} (only applicable when \code{show_wtp = TRUE})
+#' , then the shape will be mapped to that aesthetic.
+#' @param wtp_linetype The \code{linetype} used for \code{ggplot2::geom_abline()}
+#' when making the WTP. Defaults to \code{"dashed"}.
+#' @param add_prop_ce Whether to append the proportion of simulations for that
+#' method which were cost-effective (beneath the WTP threshold)
+#' to their labels in the legend. Only applicable when \code{show_wtp = TRUE}.
 #' @param ... Additional (unused) arguments.
 #'
 #' @details
@@ -100,6 +120,8 @@ ce_plot.predictNMBsim <- function(object,
                                   methods_order = NULL,
                                   rename_vector,
                                   shape = 21,
+                                  wtp_linetype = "dashed",
+                                  add_prop_ce = FALSE,
                                   ...) {
   if (missing(ref_col)) {
     stop("'ref_col' must be specified for creating a cost-effectiveness plot.")
@@ -109,7 +131,7 @@ ce_plot.predictNMBsim <- function(object,
     stop(
       "This predictNMBsim object did not track the QALYs and costs at each",
       " simulation so a cost-effectiveness plot cannot be made.",
-      " This is likley because the functions used for ",
+      " This is likely because the functions used for ",
       "'fx_nmb_training' and 'fx_nmb_evaluation' were either not made using",
       " 'get_nmb_sampler()' or, if they were, they didn't use 'qalys_lost'",
       " and 'wtp'."
@@ -139,11 +161,7 @@ ce_plot.predictNMBsim <- function(object,
     dplyr::select(-c(percentile, in_interval)) %>%
     tidyr::pivot_wider(names_from = "type", values_from = "value")
 
-  p <- p_data %>%
-    ggplot2::ggplot(ggplot2::aes(qalys, costs, col = name)) +
-    ggplot2::geom_point(shape = shape) +
-    ggplot2::geom_hline(yintercept = 0) +
-    ggplot2::geom_vline(xintercept = 0)
+  legend_aesthetics <- list(col = "Cutpoint Methods")
 
   if (show_wtp) {
     if (!is.function(attr(object$meta_data$fx_nmb_evaluation, "wtp"))) {
@@ -168,7 +186,7 @@ ce_plot.predictNMBsim <- function(object,
       )
       if (!missing(wtp)) {
         message(
-          "Using the specified wtp value to draw the cost-effectiveness plane",
+          "Using the specified wtp value to draw the WTP line",
           " and ignoring the stored wtp.\n\n"
         )
         assertthat::is.number(wtp)
@@ -186,17 +204,79 @@ ce_plot.predictNMBsim <- function(object,
       }
     }
 
+    if (add_prop_ce | shape == "cost-effective") {
+      p_data <- p_data %>%
+        dplyr::mutate(ce = costs < wtp * qalys)
+
+      if (add_prop_ce) {
+        p_data <- p_data %>%
+          dplyr::group_by(name) %>%
+          dplyr::mutate(name = paste0(name, " (", scales::percent(mean(ce)), ")"))
+
+        legend_aesthetics['col'] <- "Cutpoint Methods (cost-effective %)"
+      }
+    }
+  }
+
+  match_shape_and_col_legend <- FALSE
+  if (shape == "method") {
+    p <- p_data %>%
+      ggplot2::ggplot(ggplot2::aes(qalys, costs, col = name, shape = name)) +
+      ggplot2::geom_point() +
+      ggplot2::geom_hline(yintercept = 0) +
+      ggplot2::geom_vline(xintercept = 0)
+
+    if (length(levels(p_data$name)) > 6) {
+      p <- p +
+        ggplot2::scale_shape_manual(values = 1:length(levels(p_data$name)))
+    }
+
+    match_shape_and_col_legend <- TRUE
+
+  } else if (shape == "cost-effective") {
+    stopifnot(show_wtp)
+    p <- p_data %>%
+      dplyr::mutate(ce = costs < wtp * qalys) %>%
+      ggplot2::ggplot(ggplot2::aes(qalys, costs, col = name, shape = ce)) +
+      ggplot2::geom_point() +
+      ggplot2::geom_hline(yintercept = 0) +
+      ggplot2::geom_vline(xintercept = 0)
+
+    legend_aesthetics['shape'] <- "Cost-effective"
+
+  } else {
+    p <- p_data %>%
+      ggplot2::ggplot(ggplot2::aes(qalys, costs, col = name)) +
+      ggplot2::geom_point(shape = shape) +
+      ggplot2::geom_hline(yintercept = 0) +
+      ggplot2::geom_vline(xintercept = 0)
+  }
+
+  if (show_wtp) {
     p <- p +
       ggplot2::geom_abline(
         intercept = 0,
-        slope = wtp
+        slope = wtp,
+        linetype = wtp_linetype
       )
   }
 
-  p +
-    ggplot2::labs(
-      x = "Incremental Effectiveness (Quality-Adjusted Life Years)",
-      y = "Incremental Costs ($)",
-      col = "Cutpoint Methods"
-    )
+  if (match_shape_and_col_legend) {
+    p +
+      ggplot2::labs(
+        x = "Incremental Effectiveness (Quality-Adjusted Life Years)",
+        y = "Incremental Costs ($)",
+        col = legend_aesthetics['col'],
+        shape = legend_aesthetics['col']
+      )
+  } else {
+    p +
+      ggplot2::labs(
+        x = "Incremental Effectiveness (Quality-Adjusted Life Years)",
+        y = "Incremental Costs ($)",
+        col = legend_aesthetics['col'],
+        shape = legend_aesthetics['shape']
+      )
+  }
+
 }