+++ NEWS.md
@@ -1,3 +1,23 @@
+# predictNMB 0.2.1
+
+* `ce_plot()`
+
+  * add `wtp_linetype` argument to `ce_plot()` (and default to `"dashed"`) to 
+  given control to user and differentiate C-E plane from main axes.
+  
+  * add `shape` argument which maps to the shape aesthetic in `geom_point()`. 
+  User can also specify `shape = "cost-effective"` to map it to whether that
+  point is under the WTP/cost-effectiveness plane, or `shape = "method"` to map
+  it to the cutpoint method.
+  
+  * add `add_prop_ce` argument. If `TRUE`, the % of points for that
+  cutpoint method is concatenated to the the method name in the figure legend.
+  
+  * outcome costs were incorrectly excluded from the costs used in the 
+  incremental cost axis before. These are now added to the calculations in 
+  `evaluate_cutpoint_cost()` and reflected in the `ce_plot()`. Before, the
+  incremental costs in `ce_plot()` were treatment costs alone.
+
 # predictNMB 0.2.0
 
 * Track QALYs and intervention costs